解决后台登录问题，需要添加登录退出功能

+++ Application/Admin/Controller/UserController.class.php
@@ -3,13 +3,21 @@ namespace Admin\Controller;
 use Think\Controller;
 class UserController extends Controller {
 	public function index() {
+		$admin = session('admin');
+		if ($admin != "weread") {
+			redirect(U('User/login'), 0, "go to login");
+		}
+
 		redirect(U('User/manageUser'), 0, "redirect to manageUser");
-		// echo "hello world";
 	}
 
 
 	public function manageUser() {
-		// echo "string";
+		$admin = session('admin');
+		if ($admin != "weread") {
+			redirect(U('User/login'), 0, "go to login");
+		}
+
 		$userModel = M('users');
 		$userAll = $userModel->order('create_time', desc)->select();
 		$userCount = $userModel->count();
@@ -21,6 +29,10 @@ class UserController extends Controller {
 	}
 
 	public function deleteUser($id) {
+		$admin = session('admin');
+		if ($admin != "weread") {
+			redirect(U('User/login'), 0, "go to login");
+		}
 
 		$userModel = M('users');
 		$userModel->delete($id);
@@ -28,55 +40,52 @@ class UserController extends Controller {
 	}
 
 	public function manageBook()	{
+		$admin = session('admin');
+		if ($admin != "weread") {
+			redirect(U('User/login'), 0, "go to login");
+		}
+
 		redirect(U('Books/allBooks'), 0, "go to all books");
 	}
 
 
 	public function background() {
+		$admin = session('admin');
+		if ($admin != "weread") {
+			redirect(U('User/login'), 0, "go to login");
+		}
+
 		redirect(U('BackGround/backInfo'), 0, "go to background");
 	}
 
 	public function logout($error = 0) {
-		session('username', null);
-		session('bookid', null);
-		switch ($error) {
-			case '2':
-				redirect(U('Notes/bookAllNotes'), 0, 'go back to bookAllNotes');
-				# code...
-				break;
-			case '1':
-				redirect(U('Comments/index'), 0, "go to comments");
-			default:
-				# code...
-				break;
-		}
-		// redirect(U('Index/login'), 3, "loginout success!");
+		session('weread', null);
 		redirect(U('User/index'), 0, "loginout success!");
 
 	}
 
-	public function login($error = 0) {
+	public function login() {
 		$this.layout(true);
-		// $this->assign('error', $error);
-		// 0  --- no error
-		// 1  --- dian zan
-		// 2  --- add comment  
-		// 3  --- my bookNotes
-		// 4  --  my bookNotes
-		$this->assign('gogo', $error);
 		$this->display();
 	}
 
-	// public function login() 	{
-	// 	$this->display();
-	// }	
 	public function register() {
+		$admin = session('admin');
+		if ($admin != "weread") {
+			redirect(U('User/login'), 0, "go to login");
+		}
+
 		$this.layout(true);
 		$this->display();
 	}
 
 
 	public function addUsers() {
+		$admin = session('admin');
+		if ($admin != "weread") {
+			redirect(U('User/login'), 0, "go to login");
+		}
+
 		$user = D('users');
 		if ($user->create()) {
 			$result = $user->add();
@@ -92,41 +101,17 @@ class UserController extends Controller {
 	}
 
 
-	public function loginin($error = 0) {
+	public function loginin() {
 		$username = I('post.username');
 		$password = I('post.password');
-		$uu = (string)$username;
-		$pp = (string)$password;
-		$user = M('users');
-		$sql = "username = '$uu' and password = '$pp'";
-		$data2 = $user->where("$sql")->find();
-		if ($data2) {
-			$_SESSION['username'] = $uu;
-			// redirect(U('Index/index'), 0);
-			// redirect(U("$error"), 0);
-			switch ($error) {
-				case '1':
-				case '2':
-					redirect(U('comments/comments'), 0, "go to comments");
-					# code...
-					break;
-				case '3':
-					redirect(U('Notes/myBookNotes'), 0, "go to myBookNotes");
-					break;
-				case '4':
-					redirect(U('Notes/myAllNotes'), 0, "go to myAllNotes");
-					break;
-				default:
-					redirect(U('User/index'), 0, 'go to user');
-					# code...
-					break;
-			}
+
+		if ($username == "admin" and $password == "weread") {
+			session('admin', "weread");
+			redirect(U('BackGround/backInfo'), 0, "go to background");
 		}	else {
-				// echo "login fail";
-				$this->error("您的用户名和密码不匹配");
+			redirect(U('User/login'), 0, "go to login");
 		}
 	}
-
 
 }
